+++ test/cctest/test-heap-profiler.cc
@@ -3008,11 +3008,11 @@ class TestStatsStreamXDK : public v8::OutputStream {
     DCHECK(false);
     return kAbort;
   }
-  virtual WriteResult WriteHeapXDKChunk(const char* symbols, int symbolsSize,
-    const char* frames, int framesSize,
-    const char* types, int typesSize,
-    const char* chunks, int chunksSize,
-    const char* retentions, int retentionSize) {
+  virtual WriteResult WriteHeapXDKChunk(const char* symbols, size_t symbolsSize,
+    const char* frames, size_t framesSize,
+    const char* types, size_t typesSize,
+    const char* chunks, size_t chunksSize,
+    const char* retentions, size_t retentionSize) {
     checker_->parse(symbols, frames, types);
     chunk_ = chunks;
     return kContinue;
